requirements.txt
openai
langchain
streamlit
python-dotenv==1.0.0


